--- BetterGenshinImpact/GameTask/AutoPathing/Model/PathingTaskInfo.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BetterGenshinImpact.GameTask.AutoPathing.Model
-{
-    [Serializable]
-    public class PathingTaskInfo
-    {
-        public string Name { get; set; } = string.Empty;
-        public string Description { get; set; } = string.Empty;
-
-    }
-}
